requirements.txt
langgraph
playwright
html2text
pandas
google-generativeai
colorama
python-dotenv
